+++ CHANGELOG.md
@@ -7,6 +7,9 @@ _(Note: Smaller fixes or improvements may be omitted.)_
 Error messages are now shown on standard error stream instead of the standard
 output stream.
 
+The program will now handle files with Windows-style line endings (carriage
+return and line feed) properly.
+
 ## Version 0.5.0 (2022-11-28)
 
 The minimum required CMake version for compiling the project is raised from 2.8